+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Changed
+* **BREAKING** Banner: Vervang `icon-only` Button door Icon Button ([#3554](https://github.com/dso-toolkit/dso-toolkit/issues/3554))
+
 ### Fixed
 * Mark Bar: Uitspreken van `current` en `totalCount` bij verandering ([#3660](https://github.com/dso-toolkit/dso-toolkit/issues/3660))